ImageJ Macro for subdirectories
The imageJ macro was adopted to a headless environment and ran.
interestingly one time it gave me an error but after that the macro ran
fine, so this may be due to user pertubations on the computer when not
running headless. I removed the metadata portion as it was complicating
the script and we should be able to use native python or
python-bioformats api to extract the necessary metadata from the nd2
files and write to the respective directories just fine. Cograts, this
is a big step.