Segmentation working, producing masks
We are able to segment the phase images and produce masks for them with
a background only, cell only, and outlines_txt file. It is not insanely
computationally intense so I believe the biorep level organization
should be fine and done within hours.

Now we are moving on to applying the masks to the different channels of
images and then statistical analysis from those images to help inform
parameter choices and QA data.