Statistics beginning to work
I was able to get some of the initial statistics computed, now I am
working on the metdata assigning of Z-channel name issues to make sure
we dont make any erroneous errors with regards to the data.

To do that right now there is a metadata finder we are writing into the
same function as the file function and we either need to loop over the
parent directories for that function and assign fro mthere or we need to
make some external type of loop to look for it because I think it is
looping over the file right now. i guess technically we could also loop
over the os.directname(os.directname(file_path))) and then be able to
find per file the ome.tiff may be the best but seems like a colossal way
of assigning calculation values. Also make sure that you save the
filepath name or parentdirectory name for future referral but on top of
that very possible you could then go and map pairing using parent
directory to pait to images and reassign Channel values.