EDA has begun, working through plotting
I am currently plotting through different parameters to explore the data
and then go from there to see if any backgroun dsubtractions or changes
in imaging parameters need to be made. RIght now I am not doing this
with the full dataset based 1 on complexity but secondly omn the fact
that some of the dataset isnt parsing ccorrectly due to file naming
schema.