Reorganizing files
I am reoganizing the files in the repo, by deleting some of the files
that were made and saved along the way while troubleshooting the
pipeline as well as making subdirectories for different scripts used in
the process. As I move into data analysis I am going to start to take a
more modular approach to save functions as scripts to be able to call
down the line, i will keep the original functions in my notebook for
clarity for the reader but save them for any future use of general data
analysis for their utility.