Working Images no data warning
I think that the warnings are only for our phase images whih are being
read in as 0 since phase wasnt taken in every channel in which case ...
we wouldnt care and could leave it. But I incorporated some script to
organzie the output to make sure we dont miss something, jsut habving a
hard time printing it but can work on that another time.