Outline.txt Writing Bug Fixed
The function was changed somewhere in the process of adapting the
program to write the cell_oinly and background only and was giving us
bad outine.txt files that had nothing to do with the masks. This was a
simply fix in the omnipose_working_nostats_using_Model_abd_path.ipynb
and will need to later be incorporated into other omnipose run
notebooks.