Omnipose Retrain and Outline Fix Up to date
I finished the training script (both denovo training and adding to the
original omnipose or cellpose ML algorithms. Additionally we have our
initial stats list and background only cell only masks. So far it seems
that the higher epoch training set is leading to over segmentation but
is very good at segmenting adjacent cells with high sinuousity.

Next steps are to run the MicrobeJ outputs and analyze the data
specifically to see if our imaging parameters are leaing to bleachign
across the zstacks that is impairing out ability to interpret the data.