Adding an Environments folder
I am addint yml files to ensure that my environments stay the same
amongst different machines with the exception of the gpu computer having
its own environments since it is gpu supported. About to upgrade to the
xps 17 as well so we will see if that brings any unique challenges in
either windows or linux boot.