Initial Data Shows Z Stack Effect
I am working on creating specific cell numbers per image using a
strategy like the one below, but I can already see that there is a
zstack effect even if I cannot know for sure the datapoints are
connected (high datapoints with high datapoints for each stack etc). It
is possible that some type of bacground subtraction would enable us to
to normalize but I doubt it.