Added MicrobeJ_data_analyzed.ipynb for data checking
Added a bunch of functions to find and concatenate csv's and then
identify csvs with non matching experiment names or non-matching data
columns (meaning we didnt expand the data) there are lots of these and
having to go back and manually correct. Best to use the file paths to
navigate.